Work more on implementing/debugging flexbox

+++ src/main/java/com/github/webicitybrowser/thready/drawing/core/Canvas2D.java
@@ -39,9 +39,9 @@ public interface Canvas2D {
 	 */
 	void drawText(float x, float y, String string);
 	
-	void drawLine(float x, float y, int run, float fall);
+	void drawLine(float x, float y, float run, float fall);
 	
-	void drawTexture(float x, float y, int l, int h, Image texture);
+	void drawTexture(float x, float y, float l, float h, Image texture);
 	
 	/**
 	 * Get the paint used to draw figures.
